+++ docs/recipes/automation/tp-setup/bootstrap/utils/env.py
@@ -49,6 +49,12 @@ class EnvConfig:
     TP_AUTO_IS_PROVISION_PULSAR = os.environ.get("TP_AUTO_IS_PROVISION_PULSAR", "false").lower() == "true"
     TP_AUTO_IS_PROVISION_TIBCOHUB = os.environ.get("TP_AUTO_IS_PROVISION_TIBCOHUB", "false").lower() == "true"
 
+    # OAuth token
+    TP_AUTO_TOKEN_NAMESPACE = os.environ.get("TP_AUTO_TOKEN_NAMESPACE") or "automation"
+    TP_AUTO_TOKEN_NAME = os.environ.get("TP_AUTO_TOKEN_NAME") or "auto-token"
+    TP_AUTO_TOKEN_DURATION = os.environ.get("TP_AUTO_TOKEN_DURATION") or "3"
+    TP_AUTO_TOKEN_DURATION_UNIT = os.environ.get("TP_AUTO_TOKEN_DURATION_UNIT") or "Months"
+
     # start app or not
     TP_AUTO_START_FLOGO_APP = os.environ.get("TP_AUTO_START_FLOGO_APP", "true").lower() == "true"
     TP_AUTO_START_BWCE_APP = os.environ.get("TP_AUTO_START_BWCE_APP", "false").lower() == "true"
@@ -147,11 +153,11 @@ class EnvConfig:
     TP_AUTO_MAX_DATA_PLANE = 9
 
     # apps: bwce, bw5ce, flogo
-    BWCE_APP_FILE_NAME = os.environ.get("TP_AUTO_BWCE_APP_FILE_NAME") or "bwce-tt.ear"
+    BWCE_APP_FILE_NAME = os.environ.get("TP_AUTO_BWCE_APP_FILE_NAME") or "rest-bwce-1.ear"
     BWCE_APP_NAME = os.environ.get("BWCE_APP_NAME") or BWCE_APP_FILE_NAME.removesuffix(".ear")
     BW5CE_APP_FILE_NAME = os.environ.get("TP_AUTO_BW5CE_APP_FILE_NAME") or "bw5ce-dynamicHeaders.ear"
     BW5CE_APP_NAME = os.environ.get("BW5CE_APP_NAME") or BW5CE_APP_FILE_NAME.removesuffix(".ear")
-    FLOGO_APP_FILE_NAME = os.environ.get("TP_AUTO_FLOGO_APP_FILE_NAME") or "flogo.json"
+    FLOGO_APP_FILE_NAME = os.environ.get("TP_AUTO_FLOGO_APP_FILE_NAME") or "rest-flogo-1.json"
     # need to make sure the flogo app name is unique and lower case in the above JSON file
     FLOGO_APP_NAME = os.environ.get("FLOGO_APP_NAME") or Helper.get_app_name(FLOGO_APP_FILE_NAME)
 
@@ -168,6 +174,8 @@ class EnvConfig:
             ColorLogger.warning(f"TP_AUTO_IS_CONFIG_O11Y is false, will not config Data Plane o11y")
         if not self.TP_AUTO_IS_PROVISION_BWCE:
             ColorLogger.warning(f"TP_AUTO_IS_PROVISION_BWCE is false, will not provision BWCE capability")
+        if not self.TP_AUTO_IS_PROVISION_BW5CE:
+            ColorLogger.warning(f"TP_AUTO_IS_PROVISION_BW5CE is false, will not provision BW5CE capability")
         if not self.TP_AUTO_IS_PROVISION_EMS:
             ColorLogger.warning(f"TP_AUTO_IS_PROVISION_EMS is false, will not provision EMS capability")
         if not self.TP_AUTO_IS_PROVISION_FLOGO: